refactor(ALL): account for Nightly `vim.treesitter.get_parser` changes
On Neovim>=0.12 `vim.treesitter.get_parser()` will return `nil` if there
is no parser (instead of purely throwing error as in Neovim<=0.10 and
an error which can be disabled in favor `nil` output in Neovim=0.11).
Account for that in code to early prefer `nil` output on Neovim=0.11
(with `{ error = false }`).

Also `vim.treesitter.start()` in case of no parser found might (in the
future) silently return `nil, errmsg` in the future instead of throwing
a "No parser found" error (as throwing an error is "not documented and
is an implementation detail").
Account for that by checking first whether there is a parser available.

Resolve #1218